+++ polardbx-20200202/alibabacloud_polardbx20200202/models/_create_dbinstance_request.py
@@ -24,6 +24,7 @@ class CreateDBInstanceRequest(DaraModel):
         is_columnar_read_dbinstance: bool = None,
         is_read_dbinstance: bool = None,
         network_type: str = None,
+        origin_minor_version: str = None,
         pay_type: str = None,
         period: str = None,
         primary_dbinstance_name: str = None,
@@ -32,6 +33,7 @@ class CreateDBInstanceRequest(DaraModel):
         resource_group_id: str = None,
         secondary_zone: str = None,
         series: str = None,
+        storage_type: str = None,
         tertiary_zone: str = None,
         topology_type: str = None,
         used_time: int = None,
@@ -55,6 +57,7 @@ class CreateDBInstanceRequest(DaraModel):
         self.is_columnar_read_dbinstance = is_columnar_read_dbinstance
         self.is_read_dbinstance = is_read_dbinstance
         self.network_type = network_type
+        self.origin_minor_version = origin_minor_version
         # This parameter is required.
         self.pay_type = pay_type
         self.period = period
@@ -65,6 +68,7 @@ class CreateDBInstanceRequest(DaraModel):
         self.resource_group_id = resource_group_id
         self.secondary_zone = secondary_zone
         self.series = series
+        self.storage_type = storage_type
         self.tertiary_zone = tertiary_zone
         # This parameter is required.
         self.topology_type = topology_type
@@ -127,6 +131,9 @@ class CreateDBInstanceRequest(DaraModel):
         if self.network_type is not None:
             result['NetworkType'] = self.network_type
 
+        if self.origin_minor_version is not None:
+            result['OriginMinorVersion'] = self.origin_minor_version
+
         if self.pay_type is not None:
             result['PayType'] = self.pay_type
 
@@ -150,6 +157,9 @@ class CreateDBInstanceRequest(DaraModel):
 
         if self.series is not None:
             result['Series'] = self.series
+
+        if self.storage_type is not None:
+            result['StorageType'] = self.storage_type
 
         if self.tertiary_zone is not None:
             result['TertiaryZone'] = self.tertiary_zone
@@ -218,6 +228,9 @@ class CreateDBInstanceRequest(DaraModel):
         if m.get('NetworkType') is not None:
             self.network_type = m.get('NetworkType')
 
+        if m.get('OriginMinorVersion') is not None:
+            self.origin_minor_version = m.get('OriginMinorVersion')
+
         if m.get('PayType') is not None:
             self.pay_type = m.get('PayType')
 
@@ -241,6 +254,9 @@ class CreateDBInstanceRequest(DaraModel):
 
         if m.get('Series') is not None:
             self.series = m.get('Series')
+
+        if m.get('StorageType') is not None:
+            self.storage_type = m.get('StorageType')
 
         if m.get('TertiaryZone') is not None:
             self.tertiary_zone = m.get('TertiaryZone')